add clippy check
