+++ src/polodb_bson/value.rs
@@ -45,8 +45,7 @@ impl Value {
             (Value::Int(i1), Value::Int(i2)) => Ok(i1.cmp(i2)),
             (Value::String(str1), Value::String(str2)) => Ok(str1.cmp(str2)),
             (Value::ObjectId(oid1), Value::ObjectId(oid2)) => Ok(oid1.cmp(oid2)),
-            _ =>
-                return Err(BsonErr::TypeNotComparable(self.ty_name().into(), other.ty_name().into()))
+            _ => Err(BsonErr::TypeNotComparable(self.ty_name().into(), other.ty_name().into())),
         }
     }
