+++ src/polodb_clib/lib.rs
@@ -1,3 +1,5 @@
+#![allow(clippy::missing_safety_doc)]
+
 use polodb_core::{DbContext, DbErr, DbHandle, TransactionType, Config};
 use polodb_bson::{Value, ObjectId, Document, Array, UTCDateTime};
 use polodb_bson::linked_hash_map::Iter;
@@ -34,10 +36,8 @@ fn set_global_error(err: DbErr) {
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_open(path: *const c_char) -> *mut DbContext {
-    let cstr = unsafe {
-        CStr::from_ptr(path)
-    };
+pub unsafe extern "C" fn PLDB_open(path: *const c_char) -> *mut DbContext {
+    let cstr = CStr::from_ptr(path);
     let str = try_read_utf8!(cstr.to_str(), null_mut());
     let db = match DbContext::new(str.as_ref(), Config::default()) {
         Ok(db) => db,
@@ -51,185 +51,167 @@ pub extern "C" fn PLDB_open(path: *const c_char) -> *mut DbContext {
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_start_transaction(db: *mut DbContext, flags: c_int) -> c_int {
-    unsafe {
-        let rust_db = db.as_mut().unwrap();
-        let ty = match flags {
-            0 => None,
-            1 => Some(TransactionType::Read),
-            2 => Some(TransactionType::Write),
-            _ => {
-                set_global_error(DbErr::UnknownTransactionType);
-                return PLDB_error_code();
-            }
-        };
-        match rust_db.start_transaction(ty) {
-            Ok(()) => 0,
-            Err(err) => {
-                set_global_error(err);
-                PLDB_error_code()
-            }
+pub unsafe extern "C" fn PLDB_start_transaction(db: *mut DbContext, flags: c_int) -> c_int {
+    let rust_db = db.as_mut().unwrap();
+    let ty = match flags {
+        0 => None,
+        1 => Some(TransactionType::Read),
+        2 => Some(TransactionType::Write),
+        _ => {
+            set_global_error(DbErr::UnknownTransactionType);
+            return PLDB_error_code();
+        }
+    };
+    match rust_db.start_transaction(ty) {
+        Ok(()) => 0,
+        Err(err) => {
+            set_global_error(err);
+            PLDB_error_code()
         }
     }
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_rollback(db: *mut DbContext) -> c_int {
-    unsafe {
-        let rust_db = db.as_mut().unwrap();
-        match rust_db.rollback() {
-            Ok(()) => 0,
-            Err(err) => {
-                set_global_error(err);
-                PLDB_error_code()
-            }
+pub unsafe extern "C" fn PLDB_rollback(db: *mut DbContext) -> c_int {
+    let rust_db = db.as_mut().unwrap();
+    match rust_db.rollback() {
+        Ok(()) => 0,
+        Err(err) => {
+            set_global_error(err);
+            PLDB_error_code()
         }
     }
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_commit(db: *mut DbContext) -> c_int {
-    unsafe {
-        let rust_db = db.as_mut().unwrap();
-        match rust_db.commit() {
-            Ok(()) => 0,
-            Err(err) => {
-                set_global_error(err);
-                PLDB_error_code()
-            }
+pub unsafe extern "C" fn PLDB_commit(db: *mut DbContext) -> c_int {
+    let rust_db = db.as_mut().unwrap();
+    match rust_db.commit() {
+        Ok(()) => 0,
+        Err(err) => {
+            set_global_error(err);
+            PLDB_error_code()
         }
     }
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_count(db: *mut DbContext, col_id: c_uint, meta_version: u32) -> c_longlong {
-    unsafe {
-        let rust_db = db.as_mut().unwrap();
-        let result = rust_db.count(col_id, meta_version);
-        match result {
-            Ok(result) => {
-                return result as c_longlong;
-            }
-            Err(err) => {
-                set_global_error(err);
-                return PLDB_error_code() as c_longlong;
-            }
-        };
-    }
+pub unsafe extern "C" fn PLDB_count(db: *mut DbContext, col_id: c_uint, meta_version: u32) -> c_longlong {
+    let rust_db = db.as_mut().unwrap();
+    let result = rust_db.count(col_id, meta_version);
+    match result {
+        Ok(result) => {
+            return result as c_longlong;
+        }
+        Err(err) => {
+            set_global_error(err);
+            return PLDB_error_code() as c_longlong;
+        }
+    };
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_create_collection(db: *mut DbContext,
+pub unsafe extern "C" fn PLDB_create_collection(db: *mut DbContext,
                                          name: *const c_char,
                                          col_id: *mut c_uint,
                                          meta_version: *mut c_uint) -> c_int {
-    unsafe {
-        let name_str= CStr::from_ptr(name);
-        let name_utf8 = try_read_utf8!(name_str.to_str(), PLDB_error_code());
-        let oid_result = db.as_mut().unwrap().create_collection(name_utf8);
-        match oid_result {
-            Ok(meta) => {
-                col_id.write(meta.id);
-                meta_version.write(meta.meta_version);
-                0
-            }
-
-            Err(err) => {
-                set_global_error(err);
-                PLDB_error_code()
-            }
-
+    let name_str= CStr::from_ptr(name);
+    let name_utf8 = try_read_utf8!(name_str.to_str(), PLDB_error_code());
+    let oid_result = db.as_mut().unwrap().create_collection(name_utf8);
+    match oid_result {
+        Ok(meta) => {
+            col_id.write(meta.id);
+            meta_version.write(meta.meta_version);
+            0
         }
+
+        Err(err) => {
+            set_global_error(err);
+            PLDB_error_code()
+        }
+
     }
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_drop(db: *mut DbContext, col_id: c_uint, meta_version: c_uint) -> c_int {
-    unsafe {
-        let result = db.as_mut().unwrap().drop(col_id, meta_version);
-        if let Err(err) = result {
-            set_global_error(err);
-            return PLDB_error_code();
-        }
+pub unsafe extern "C" fn PLDB_drop(db: *mut DbContext, col_id: c_uint, meta_version: c_uint) -> c_int {
+    let result = db.as_mut().unwrap().drop(col_id, meta_version);
+    if let Err(err) = result {
+        set_global_error(err);
+        return PLDB_error_code();
     }
     0
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_get_collection_meta_by_name(db: *mut DbContext, name: *const c_char, id: *mut c_uint, version: *mut c_uint) -> c_int {
-    unsafe {
-        let str = CStr::from_ptr(name);
-        let utf8str = try_read_utf8!(str.to_str(), PLDB_error_code());
-        let result = db.as_mut().unwrap().get_collection_meta_by_name(utf8str);
-        return match result {
-            Ok(info) => {
-                id.write(info.id);
-                version.write(info.meta_version);
-                0
-            }
+pub unsafe extern "C" fn PLDB_get_collection_meta_by_name(db: *mut DbContext, name: *const c_char, id: *mut c_uint, version: *mut c_uint) -> c_int {
+    let str = CStr::from_ptr(name);
+    let utf8str = try_read_utf8!(str.to_str(), PLDB_error_code());
+    let result = db.as_mut().unwrap().get_collection_meta_by_name(utf8str);
+    return match result {
+        Ok(info) => {
+            id.write(info.id);
+            version.write(info.meta_version);
+            0
+        }
 
-            Err(err) => {
-                set_global_error(err);
-                PLDB_error_code()
-            }
+        Err(err) => {
+            set_global_error(err);
+            PLDB_error_code()
         }
     }
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_insert(db: *mut DbContext, col_id: c_uint, meta_version: c_uint, doc: *const Rc<Document>) -> c_int {
-    unsafe {
-        let local_db = db.as_mut().unwrap();
-        let local_doc = doc.as_ref().unwrap().clone();
-        let insert_result = local_db.insert(col_id, meta_version, local_doc);
-        if let Err(err) = insert_result {
-            set_global_error(err);
-            return PLDB_error_code();
-        }
+pub unsafe extern "C" fn PLDB_insert(db: *mut DbContext, col_id: c_uint, meta_version: c_uint, doc: *const Rc<Document>) -> c_int {
+    let local_db = db.as_mut().unwrap();
+    let local_doc = doc.as_ref().unwrap().clone();
+    let insert_result = local_db.insert(col_id, meta_version, local_doc);
+    if let Err(err) = insert_result {
+        set_global_error(err);
+        return PLDB_error_code();
     }
     0
 }
 
 /// query is nullable
 #[no_mangle]
-pub extern "C" fn PLDB_find(db: *mut DbContext,
+pub unsafe extern "C" fn PLDB_find(db: *mut DbContext,
                             col_id: c_uint,
                             meta_version: c_uint,
                             query: *const Rc<Document>,
                             out_handle: *mut *mut DbHandle) -> c_int {
-    unsafe {
-        let rust_db = db.as_mut().unwrap();
+    let rust_db = db.as_mut().unwrap();
 
-        let handle_result = match query.as_ref() {
-            Some(query_doc) => rust_db.find(col_id, meta_version, Some(query_doc.borrow())),
-            None => rust_db.find(col_id, meta_version, None),
-        };
+    let handle_result = match query.as_ref() {
+        Some(query_doc) => rust_db.find(col_id, meta_version, Some(query_doc.borrow())),
+        None => rust_db.find(col_id, meta_version, None),
+    };
 
-        let handle = match handle_result {
-            Ok(handle) => handle,
-            Err(err) => {
-                set_global_error(err);
-                return PLDB_error_code();
-            }
-        };
+    let handle = match handle_result {
+        Ok(handle) => handle,
+        Err(err) => {
+            set_global_error(err);
+            return PLDB_error_code();
+        }
+    };
 
-        let boxed_handle = Box::new(handle);
-        let raw_handle = Box::into_raw(boxed_handle);
+    let boxed_handle = Box::new(handle);
+    let raw_handle = Box::into_raw(boxed_handle);
 
-        out_handle.write(raw_handle);
-    }
+    out_handle.write(raw_handle);
 
     0
 }
 
 /// query is nullable
 #[no_mangle]
-pub extern "C" fn PLDB_update(db: *mut DbContext,
+pub unsafe extern "C" fn PLDB_update(db: *mut DbContext,
                               col_id: c_uint,
                               meta_version: c_uint,
                               query: *const Rc<Document>,
                               update: *const Rc<Document>) -> c_longlong {
-    let result = unsafe {
+    let result = {
         let rust_db = db.as_mut().unwrap();
 
         let update_doc = update.as_ref().unwrap();
@@ -251,8 +233,8 @@ pub extern "C" fn PLDB_update(db: *mut DbContext,
 
 /// return value represents how many rows are deleted
 #[no_mangle]
-pub extern "C" fn PLDB_delete(db: *mut DbContext, col_id: c_uint, meta_version: c_uint, query: *const Rc<Document>) -> c_longlong {
-    let result = unsafe {
+pub unsafe extern "C" fn PLDB_delete(db: *mut DbContext, col_id: c_uint, meta_version: c_uint, query: *const Rc<Document>) -> c_longlong {
+    let result = {
         let rust_db = db.as_mut().unwrap();
 
         let query_doc = query.as_ref().unwrap();
@@ -286,66 +268,56 @@ pub extern "C" fn PLDB_delete_all(db: *mut DbContext, col_id: c_uint, meta_versi
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_handle_to_str(handle: *mut DbHandle, buffer: *mut c_char, buffer_size: c_uint) -> c_int {
-    unsafe {
-        let rust_handle = handle.as_mut().unwrap();
-        let str_content = format!("{}", rust_handle);
-        let length = str_content.len();
+pub unsafe extern "C" fn PLDB_handle_to_str(handle: *mut DbHandle, buffer: *mut c_char, buffer_size: c_uint) -> c_int {
+    let rust_handle = handle.as_mut().unwrap();
+    let str_content = format!("{}", rust_handle);
+    let length = str_content.len();
 
-        if buffer.is_null() {
-            return (length + 1) as c_int;
-        }
-
-        if (buffer_size as usize) < length + 1 {
-            set_global_error(DbErr::BufferNotEnough(length + 1));
-            return PLDB_error_code();
-        }
-
-        let cstring = CString::new(str_content).unwrap();
-        cstring.as_ptr().copy_to_nonoverlapping(buffer, length);
-
-        length as c_int
+    if buffer.is_null() {
+        return (length + 1) as c_int;
     }
+
+    if (buffer_size as usize) < length + 1 {
+        set_global_error(DbErr::BufferNotEnough(length + 1));
+        return PLDB_error_code();
+    }
+
+    let cstring = CString::new(str_content).unwrap();
+    cstring.as_ptr().copy_to_nonoverlapping(buffer, length);
+
+    length as c_int
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_handle_step(handle: *mut DbHandle) -> c_int {
-    unsafe {
-        let rust_handle = handle.as_mut().unwrap();
-        let result = rust_handle.step();
+pub unsafe extern "C" fn PLDB_handle_step(handle: *mut DbHandle) -> c_int {
+    let rust_handle = handle.as_mut().unwrap();
+    let result = rust_handle.step();
 
-        if let Err(err) = result {
-            set_global_error(err);
-            return PLDB_error_code();
-        }
-
-        0
+    if let Err(err) = result {
+        set_global_error(err);
+        return PLDB_error_code();
     }
+
+    0
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_handle_state(handle: *mut DbHandle) -> c_int {
-    unsafe {
-        let rust_handle = handle.as_mut().unwrap();
-        rust_handle.state() as c_int
-    }
+pub unsafe extern "C" fn PLDB_handle_state(handle: *mut DbHandle) -> c_int {
+    let rust_handle = handle.as_mut().unwrap();
+    rust_handle.state() as c_int
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_handle_get(handle: *mut DbHandle, out_val: *mut *mut Value) {
-    unsafe {
-        let rust_handle = handle.as_mut().unwrap();
-        let boxed_handle = Box::new(rust_handle.get().clone());
-        let handle_ptr = Box::into_raw(boxed_handle);
-        out_val.write(handle_ptr);
-    }
+pub unsafe extern "C" fn PLDB_handle_get(handle: *mut DbHandle, out_val: *mut *mut Value) {
+    let rust_handle = handle.as_mut().unwrap();
+    let boxed_handle = Box::new(rust_handle.get().clone());
+    let handle_ptr = Box::into_raw(boxed_handle);
+    out_val.write(handle_ptr);
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_free_handle(handle: *mut DbHandle) {
-    unsafe {
-        let _ptr = Box::from_raw(handle);
-    }
+pub unsafe extern "C" fn PLDB_free_handle(handle: *mut DbHandle) {
+    let _ptr = Box::from_raw(handle);
 }
 
 #[no_mangle]
@@ -382,21 +354,19 @@ pub extern "C" fn PLDB_error_msg() -> *const c_char {
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_version(buffer: *mut c_char, buffer_size: c_uint) -> c_uint {
+pub unsafe extern "C" fn PLDB_version(buffer: *mut c_char, buffer_size: c_uint) -> c_uint {
     let version_str = DbContext::get_version();
     let str_size = version_str.len();
     let cstring = CString::new(version_str).unwrap();
-    unsafe {
-        let c_ptr = cstring.as_ptr();
-        let expected_size: usize = std::cmp::min(str_size, buffer_size as usize);
-        c_ptr.copy_to(buffer, expected_size);
-        expected_size as c_uint
-    }
+    let c_ptr = cstring.as_ptr();
+    let expected_size: usize = std::cmp::min(str_size, buffer_size as usize);
+    c_ptr.copy_to(buffer, expected_size);
+    expected_size as c_uint
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_close(db: *mut DbContext) {
-    let _ptr = unsafe { Box::from_raw(db) };
+pub unsafe extern "C" fn PLDB_close(db: *mut DbContext) {
+    let _ptr = Box::from_raw(db);
 }
 
 #[no_mangle]
@@ -406,149 +376,131 @@ pub extern "C" fn PLDB_mk_int(val: i64) -> *mut Value {
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_value_type(val: *const Value) -> c_int {
-    unsafe {
-        let local_val = val.as_ref().unwrap();
-        let ty = local_val.ty_int();
+pub unsafe extern "C" fn PLDB_value_type(val: *const Value) -> c_int {
+    let local_val = val.as_ref().unwrap();
+    let ty = local_val.ty_int();
 
-        ty as c_int
+    ty as c_int
+}
+
+#[no_mangle]
+pub unsafe extern "C" fn PLDB_value_get_i64(val: *const Value, out_val: *mut i64) -> c_int {
+    let local_val = val.as_ref().unwrap();
+    match local_val {
+        Value::Int(i) => {
+            out_val.write(*i);
+            0
+        }
+
+        _ => -1
+
     }
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_value_get_i64(val: *const Value, out_val: *mut i64) -> c_int {
-    unsafe {
-        let local_val = val.as_ref().unwrap();
-        match local_val {
-            Value::Int(i) => {
-                out_val.write(*i);
+pub unsafe extern "C" fn PLDB_value_get_bool(val: *const Value) -> c_int {
+    let local_val = val.as_ref().unwrap();
+    match local_val {
+        Value::Boolean(bl) =>{
+            if *bl {
+                1
+            } else {
                 0
             }
-
-            _ => -1
-
         }
+
+        _ => -1,
+
     }
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_value_get_bool(val: *const Value) -> c_int {
-    unsafe {
-        let local_val = val.as_ref().unwrap();
-        match local_val {
-            Value::Boolean(bl) =>{
-                if *bl {
-                    1
-                } else {
-                    0
-                }
-            }
-
-            _ => -1,
-
+pub unsafe extern "C" fn PLDB_value_get_double(val: *const Value, out: *mut c_double) -> c_int {
+    let local_val = val.as_ref().unwrap();
+    match local_val {
+        Value::Double(num) => {
+            out.write(*num);
+            0
         }
+
+        _ => -1,
+
     }
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_value_get_double(val: *const Value, out: *mut c_double) -> c_int {
-    unsafe {
-        let local_val = val.as_ref().unwrap();
-        match local_val {
-            Value::Double(num) => {
-                out.write(*num);
-                0
-            }
-
-            _ => -1,
-
+pub unsafe extern "C" fn PLDB_value_get_array(val: *const Value, out: *mut *mut Rc<Array>) -> c_int {
+    let local_val = val.as_ref().unwrap();
+    match local_val {
+        Value::Array(arr) => {
+            let boxed_array = Box::new(arr.clone());
+            out.write(Box::into_raw(boxed_array));
+            0
         }
+
+        _ => -1,
+
     }
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_value_get_array(val: *const Value, out: *mut *mut Rc<Array>) -> c_int {
-    unsafe {
-        let local_val = val.as_ref().unwrap();
-        match local_val {
-            Value::Array(arr) => {
-                let boxed_array = Box::new(arr.clone());
-                out.write(Box::into_raw(boxed_array));
-                0
-            }
-
-            _ => -1,
-
+pub unsafe extern "C" fn PLDB_value_get_object_id(val: *const Value, out: *mut *mut ObjectId) -> c_int {
+    let local_val = val.as_ref().unwrap();
+    match local_val {
+        Value::ObjectId(oid) => {
+            let boxed_oid: Box<ObjectId> = Box::new(oid.as_ref().clone());
+            out.write(Box::into_raw(boxed_oid));
+            0
         }
+
+        _ => -1,
+
     }
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_value_get_object_id(val: *const Value, out: *mut *mut ObjectId) -> c_int {
-    unsafe {
-        let local_val = val.as_ref().unwrap();
-        match local_val {
-            Value::ObjectId(oid) => {
-                let boxed_oid: Box<ObjectId> = Box::new(oid.as_ref().clone());
-                out.write(Box::into_raw(boxed_oid));
-                0
-            }
-
-            _ => -1,
-
+pub unsafe extern "C" fn PLDB_value_get_document(val: *const Value, out: *mut *mut Rc<Document>) -> c_int {
+    let local_val = val.as_ref().unwrap();
+    match local_val {
+        Value::Document(doc) => {
+            let boxed_doc = Box::new(doc.clone());
+            out.write(Box::into_raw(boxed_doc));
+            0
         }
+
+        _ => -1,
+
     }
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_value_get_document(val: *const Value, out: *mut *mut Rc<Document>) -> c_int {
-    unsafe {
-        let local_val = val.as_ref().unwrap();
-        match local_val {
-            Value::Document(doc) => {
-                let boxed_doc = Box::new(doc.clone());
-                out.write(Box::into_raw(boxed_doc));
-                0
-            }
+pub unsafe extern "C" fn PLDB_value_get_string_utf8(val: *const Value, out_str: *mut *const c_char) -> c_int {
+    let local_val = val.as_ref().unwrap();
+    match local_val {
+        Value::String(str) => {
+            let len = str.len();
+            let str_ptr = str.as_ptr().cast::<c_char>();
 
-            _ => -1,
+            out_str.write(str_ptr);
 
+            len as c_int
         }
+
+        _ => -1,
     }
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_value_get_string_utf8(val: *const Value, out_str: *mut *const c_char) -> c_int {
-    unsafe {
-        let local_val = val.as_ref().unwrap();
-        match local_val {
-            Value::String(str) => {
-                let len = str.len();
-                let str_ptr = str.as_ptr().cast::<c_char>();
-
-                out_str.write(str_ptr);
-
-                len as c_int
-            }
-
-            _ => -1,
+pub unsafe extern "C" fn PLDB_value_get_utc_datetime(val: *const Value, out_time: *mut *mut UTCDateTime) -> c_int {
+    let local_val = val.as_ref().unwrap();
+    match local_val {
+        Value::UTCDateTime(dt) => {
+            let boxed_time = Box::new(dt.as_ref().clone());
+            out_time.write(Box::into_raw(boxed_time));
+            0
         }
-    }
-}
 
-#[no_mangle]
-pub extern "C" fn PLDB_value_get_utc_datetime(val: *const Value, out_time: *mut *mut UTCDateTime) -> c_int {
-    unsafe {
-        let local_val = val.as_ref().unwrap();
-        match local_val {
-            Value::UTCDateTime(dt) => {
-                let boxed_time = Box::new(dt.as_ref().clone());
-                out_time.write(Box::into_raw(boxed_time));
-                0
-            }
-
-            _ => -1,
-        }
+        _ => -1,
     }
 }
 
@@ -565,146 +517,119 @@ pub extern "C" fn PLDB_mk_arr_with_size(size: c_uint) -> *mut Rc<Array> {
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_free_arr(arr: *mut Rc<Array>) {
-    let _ptr = unsafe { Box::from_raw(arr) };
+pub unsafe extern "C" fn PLDB_free_arr(arr: *mut Rc<Array>) {
+    let _ptr = Box::from_raw(arr);
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_arr_push(arr: *mut Rc<Array>, val: *const Value) {
-    unsafe {
-        let local_arr = arr.as_mut().unwrap();
-        let arr_mut = Rc::get_mut(local_arr).unwrap();
-        let local_val = val.as_ref().unwrap();
-        arr_mut.push(local_val.clone())
-    }
+pub unsafe extern "C" fn PLDB_arr_push(arr: *mut Rc<Array>, val: *const Value) {
+    let local_arr = arr.as_mut().unwrap();
+    let arr_mut = Rc::get_mut(local_arr).unwrap();
+    let local_val = val.as_ref().unwrap();
+    arr_mut.push(local_val.clone())
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_arr_set_null(doc: *mut Rc<Array>, index: c_uint) -> c_int {
-    unsafe {
-        let local_arr = doc.as_mut().unwrap();
-        let local_arr_mut = Rc::get_mut(local_arr).unwrap();
-        local_arr_mut[index as usize] = Value::Null;
-    }
+pub unsafe extern "C" fn PLDB_arr_set_null(doc: *mut Rc<Array>, index: c_uint) -> c_int {
+    let local_arr = doc.as_mut().unwrap();
+    let local_arr_mut = Rc::get_mut(local_arr).unwrap();
+    local_arr_mut[index as usize] = Value::Null;
     0
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_arr_set_int(arr: *mut Rc<Array>, index: c_uint, value: i64) -> c_int {
-    unsafe {
-        let local_arr = arr.as_mut().unwrap();
-        let local_arr_mut = Rc::get_mut(local_arr).unwrap();
-        local_arr_mut[index as usize] = value.into();
-    }
+pub unsafe extern "C" fn PLDB_arr_set_int(arr: *mut Rc<Array>, index: c_uint, value: i64) -> c_int {
+    let local_arr = arr.as_mut().unwrap();
+    let local_arr_mut = Rc::get_mut(local_arr).unwrap();
+    local_arr_mut[index as usize] = value.into();
     0
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_arr_set_bool(arr: *mut Rc<Array>, index: c_uint, value: c_int) -> c_int {
-    unsafe {
-        let local_arr = arr.as_mut().unwrap();
-        let local_arr_mut = Rc::get_mut(local_arr).unwrap();
-        local_arr_mut[index as usize] = Value::Boolean(if value == 0 { false } else { true });
-    }
+pub unsafe extern "C" fn PLDB_arr_set_bool(arr: *mut Rc<Array>, index: c_uint, value: c_int) -> c_int {
+    let local_arr = arr.as_mut().unwrap();
+    let local_arr_mut = Rc::get_mut(local_arr).unwrap();
+    local_arr_mut[index as usize] = Value::Boolean(if value == 0 { false } else { true });
     0
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_arr_set_double(arr: *mut Rc<Array>, index: c_uint, value: f64) -> c_int {
-    unsafe {
-        let local_arr = arr.as_mut().unwrap();
-        let local_arr_mut = Rc::get_mut(local_arr).unwrap();
-        local_arr_mut[index as usize] = value.into();
-    }
+pub unsafe extern "C" fn PLDB_arr_set_double(arr: *mut Rc<Array>, index: c_uint, value: f64) -> c_int {
+    let local_arr = arr.as_mut().unwrap();
+    let local_arr_mut = Rc::get_mut(local_arr).unwrap();
+    local_arr_mut[index as usize] = value.into();
     0
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_arr_set_string(arr: *mut Rc<Array>, index: c_uint, value: *const c_char) -> c_int {
-    unsafe {
-        let local_arr = arr.as_mut().unwrap();
-        let value_str = CStr::from_ptr(value);
-        let utf8_value = try_read_utf8!(value_str.to_str(), PLDB_error_code());
-        let local_arr_mut = Rc::get_mut(local_arr).unwrap();
-        local_arr_mut[index as usize] = utf8_value.into();
-    }
+pub unsafe extern "C" fn PLDB_arr_set_string(arr: *mut Rc<Array>, index: c_uint, value: *const c_char) -> c_int {
+    let local_arr = arr.as_mut().unwrap();
+    let value_str = CStr::from_ptr(value);
+    let utf8_value = try_read_utf8!(value_str.to_str(), PLDB_error_code());
+    let local_arr_mut = Rc::get_mut(local_arr).unwrap();
+    local_arr_mut[index as usize] = utf8_value.into();
     0
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_arr_set_binary(arr: *mut Rc<Array>, index: c_uint, data: *mut c_uchar, size: c_uint) -> c_int {
+pub unsafe extern "C" fn PLDB_arr_set_binary(arr: *mut Rc<Array>, index: c_uint, data: *mut c_uchar, size: c_uint) -> c_int {
     let mut buffer: Vec<u8> = Vec::with_capacity(size as usize);
     buffer.resize(size as usize, 0);
-    unsafe {
-        data.copy_to(buffer.as_mut_ptr(), size as usize);
-        let local_arr = arr.as_mut().unwrap();
-        let local_arr_mut = Rc::get_mut(local_arr).unwrap();
-        local_arr_mut[index as usize] = buffer.into();
-    }
+    data.copy_to(buffer.as_mut_ptr(), size as usize);
+    let local_arr = arr.as_mut().unwrap();
+    let local_arr_mut = Rc::get_mut(local_arr).unwrap();
+    local_arr_mut[index as usize] = buffer.into();
     0
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_arr_set_arr(arr: *mut Rc<Array>, index: c_uint, value: *const Rc<Array>) -> c_int {
-    unsafe {
-        let local_arr = arr.as_mut().unwrap();
-        let local_arr_mut = Rc::get_mut(local_arr).unwrap();
-        let value_arr = value.as_ref().unwrap();
-        local_arr_mut[index as usize] = Value::Array(value_arr.clone());
-    }
+pub unsafe extern "C" fn PLDB_arr_set_arr(arr: *mut Rc<Array>, index: c_uint, value: *const Rc<Array>) -> c_int {
+    let local_arr = arr.as_mut().unwrap();
+    let local_arr_mut = Rc::get_mut(local_arr).unwrap();
+    let value_arr = value.as_ref().unwrap();
+    local_arr_mut[index as usize] = Value::Array(value_arr.clone());
     0
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_arr_set_doc(arr: *mut Rc<Array>, index: c_uint, value: *const Rc<Document>) -> c_uint {
-    unsafe {
-        let local_arr = arr.as_mut().unwrap();
-        let local_arr_mut = Rc::get_mut(local_arr).unwrap();
-        let value_doc = value.as_ref().unwrap();
-        local_arr_mut[index as usize] = Value::Document(value_doc.clone());
-    }
+pub unsafe extern "C" fn PLDB_arr_set_doc(arr: *mut Rc<Array>, index: c_uint, value: *const Rc<Document>) -> c_uint {
+    let local_arr = arr.as_mut().unwrap();
+    let local_arr_mut = Rc::get_mut(local_arr).unwrap();
+    let value_doc = value.as_ref().unwrap();
+    local_arr_mut[index as usize] = Value::Document(value_doc.clone());
     0
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_arr_set_UTCDateTime(arr: *mut Rc<Array>, index: c_uint, value: i64) -> c_uint {
-    unsafe {
-        let local_arr = arr.as_mut().unwrap();
-        let local_arr_mut = Rc::get_mut(local_arr).unwrap();
-        local_arr_mut[index as usize] = Value::UTCDateTime(Rc::new(UTCDateTime::new(value as u64)));
-    }
+pub unsafe extern "C" fn PLDB_arr_set_UTCDateTime(arr: *mut Rc<Array>, index: c_uint, value: i64) -> c_uint {
+    let local_arr = arr.as_mut().unwrap();
+    let local_arr_mut = Rc::get_mut(local_arr).unwrap();
+    local_arr_mut[index as usize] = Value::UTCDateTime(Rc::new(UTCDateTime::new(value as u64)));
     0
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_arr_set_object_id(arr: *mut Rc<Array>, index: c_uint, value: *const ObjectId) -> c_uint {
-    unsafe {
-        let local_arr = arr.as_mut().unwrap();
-        let local_arr_mut = Rc::get_mut(local_arr).unwrap();
-        let value_oid = value.as_ref().unwrap();
-        local_arr_mut[index as usize] = Value::ObjectId(Rc::new(value_oid.clone()));
-    }
+pub unsafe extern "C" fn PLDB_arr_set_object_id(arr: *mut Rc<Array>, index: c_uint, value: *const ObjectId) -> c_uint {
+    let local_arr = arr.as_mut().unwrap();
+    let local_arr_mut = Rc::get_mut(local_arr).unwrap();
+    let value_oid = value.as_ref().unwrap();
+    local_arr_mut[index as usize] = Value::ObjectId(Rc::new(value_oid.clone()));
     0
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_arr_get(arr: *mut Rc<Array>, index: c_uint, out_val: *mut *mut Value) -> c_int {
-    unsafe {
-        let local_arr = arr.as_mut().unwrap();
-        let val = &local_arr[index as usize];
-        let out_box = Box::new(val.clone());
-        out_val.write(Box::into_raw(out_box));
-        0
-    }
-
+pub unsafe extern "C" fn PLDB_arr_get(arr: *mut Rc<Array>, index: c_uint, out_val: *mut *mut Value) -> c_int {
+    let local_arr = arr.as_mut().unwrap();
+    let val = &local_arr[index as usize];
+    let out_box = Box::new(val.clone());
+    out_val.write(Box::into_raw(out_box));
+    0
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_arr_len(arr: *mut Rc<Array>) -> c_uint {
-    unsafe {
-        let local_arr = arr.as_ref().unwrap();
-        local_arr.len()
-    }
+pub unsafe extern "C" fn PLDB_arr_len(arr: *mut Rc<Array>) -> c_uint {
+    let local_arr = arr.as_ref().unwrap();
+    local_arr.len()
 }
 
 #[no_mangle]
@@ -714,232 +639,198 @@ pub extern "C" fn PLDB_mk_doc() -> *mut Rc<Document> {
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_doc_set(doc: *mut Rc<Document>, key: *const c_char, value: *const Value) -> c_int {
-    unsafe {
-        let local_doc = doc.as_mut().unwrap();
-        let key_str = CStr::from_ptr(key);
-        let local_value = value.as_ref().unwrap();
-        let key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
-        let local_doc_mut = Rc::get_mut(local_doc).unwrap();
-        let result = local_doc_mut.insert(key.to_string(), local_value.clone());
-        if let Some(_) = result {
-            1
-        } else {
-            0
-        }
+pub unsafe extern "C" fn PLDB_doc_set(doc: *mut Rc<Document>, key: *const c_char, value: *const Value) -> c_int {
+    let local_doc = doc.as_mut().unwrap();
+    let key_str = CStr::from_ptr(key);
+    let local_value = value.as_ref().unwrap();
+    let key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
+    let local_doc_mut = Rc::get_mut(local_doc).unwrap();
+    let result = local_doc_mut.insert(key.to_string(), local_value.clone());
+    if let Some(_) = result {
+        1
+    } else {
+        0
     }
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_doc_set_null(doc: *mut Rc<Document>, key: *const c_char) -> c_int {
-    unsafe {
-        let local_doc = doc.as_mut().unwrap();
-        let key_str = CStr::from_ptr(key);
-        let utf8_key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
-        let local_doc_mut = Rc::get_mut(local_doc).unwrap();
-        local_doc_mut.insert(utf8_key.into(), Value::Null);
-    }
+pub unsafe extern "C" fn PLDB_doc_set_null(doc: *mut Rc<Document>, key: *const c_char) -> c_int {
+    let local_doc = doc.as_mut().unwrap();
+    let key_str = CStr::from_ptr(key);
+    let utf8_key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
+    let local_doc_mut = Rc::get_mut(local_doc).unwrap();
+    local_doc_mut.insert(utf8_key.into(), Value::Null);
     0
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_doc_set_int(doc: *mut Rc<Document>, key: *const c_char, value: i64) -> c_int {
-    unsafe {
-        let local_doc = doc.as_mut().unwrap();
-        let key_str = CStr::from_ptr(key);
-        let utf8_key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
-        let local_doc_mut = Rc::get_mut(local_doc).unwrap();
-        local_doc_mut.insert(utf8_key.into(), value.into());
-    }
+pub unsafe extern "C" fn PLDB_doc_set_int(doc: *mut Rc<Document>, key: *const c_char, value: i64) -> c_int {
+    let local_doc = doc.as_mut().unwrap();
+    let key_str = CStr::from_ptr(key);
+    let utf8_key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
+    let local_doc_mut = Rc::get_mut(local_doc).unwrap();
+    local_doc_mut.insert(utf8_key.into(), value.into());
     0
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_doc_set_bool(doc: *mut Rc<Document>, key: *const c_char, value: c_int) -> c_int {
-    unsafe {
-        let local_doc = doc.as_mut().unwrap();
-        let key_str = CStr::from_ptr(key);
-        let utf8_key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
-        let local_doc_mut = Rc::get_mut(local_doc).unwrap();
-        local_doc_mut.insert(utf8_key.into(), Value::Boolean(if value == 0 { false } else { true }));
-    }
+pub unsafe extern "C" fn PLDB_doc_set_bool(doc: *mut Rc<Document>, key: *const c_char, value: c_int) -> c_int {
+    let local_doc = doc.as_mut().unwrap();
+    let key_str = CStr::from_ptr(key);
+    let utf8_key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
+    let local_doc_mut = Rc::get_mut(local_doc).unwrap();
+    local_doc_mut.insert(utf8_key.into(), Value::Boolean(value != 0));
     0
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_doc_set_double(doc: *mut Rc<Document>, key: *const c_char, value: f64) -> c_int {
-    unsafe {
-        let local_doc = doc.as_mut().unwrap();
-        let key_str = CStr::from_ptr(key);
-        let utf8_key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
-        let local_doc_mut = Rc::get_mut(local_doc).unwrap();
-        local_doc_mut.insert(utf8_key.into(), value.into());
-    }
+pub unsafe extern "C" fn PLDB_doc_set_double(doc: *mut Rc<Document>, key: *const c_char, value: f64) -> c_int {
+    let local_doc = doc.as_mut().unwrap();
+    let key_str = CStr::from_ptr(key);
+    let utf8_key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
+    let local_doc_mut = Rc::get_mut(local_doc).unwrap();
+    local_doc_mut.insert(utf8_key.into(), value.into());
     0
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_doc_set_string(doc: *mut Rc<Document>, key: *const c_char, value: *const c_char) -> c_int {
-    unsafe {
-        let local_doc = doc.as_mut().unwrap();
-        let key_str = CStr::from_ptr(key);
-        let utf8_key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
-        let value_str = CStr::from_ptr(value);
-        let utf8_value = try_read_utf8!(value_str.to_str(), PLDB_error_code());
-        let local_doc_mut = Rc::get_mut(local_doc).unwrap();
-        local_doc_mut.insert(utf8_key.into(), utf8_value.into());
-    }
+pub unsafe extern "C" fn PLDB_doc_set_string(doc: *mut Rc<Document>, key: *const c_char, value: *const c_char) -> c_int {
+    let local_doc = doc.as_mut().unwrap();
+    let key_str = CStr::from_ptr(key);
+    let utf8_key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
+    let value_str = CStr::from_ptr(value);
+    let utf8_value = try_read_utf8!(value_str.to_str(), PLDB_error_code());
+    let local_doc_mut = Rc::get_mut(local_doc).unwrap();
+    local_doc_mut.insert(utf8_key.into(), utf8_value.into());
     0
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_doc_set_binary(doc: *mut Rc<Document>, key: *const c_char, data: *mut c_uchar, size: c_uint) -> c_int {
+pub unsafe extern "C" fn PLDB_doc_set_binary(doc: *mut Rc<Document>, key: *const c_char, data: *mut c_uchar, size: c_uint) -> c_int {
     let mut buffer: Vec<u8> = Vec::with_capacity(size as usize);
     buffer.resize(size as usize, 0);
-    unsafe {
-        data.copy_to(buffer.as_mut_ptr(), size as usize);
-        let local_doc = doc.as_mut().unwrap();
-        let key_str = CStr::from_ptr(key);
-        let utf8_key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
-        let local_doc_mut = Rc::get_mut(local_doc).unwrap();
-        local_doc_mut.insert(utf8_key.into(), buffer.into());
-    }
+    data.copy_to(buffer.as_mut_ptr(), size as usize);
+    let local_doc = doc.as_mut().unwrap();
+    let key_str = CStr::from_ptr(key);
+    let utf8_key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
+    let local_doc_mut = Rc::get_mut(local_doc).unwrap();
+    local_doc_mut.insert(utf8_key.into(), buffer.into());
     0
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_doc_set_doc(doc: *mut Rc<Document>, key: *const c_char, value: *const Rc<Document>) -> c_int {
-    unsafe {
-        let local_doc = doc.as_mut().unwrap();
-        let key_str = CStr::from_ptr(key);
-        let utf8_key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
-        let local_doc_mut = Rc::get_mut(local_doc).unwrap();
-        let value_doc = value.as_ref().unwrap();
-        local_doc_mut.insert(utf8_key.into(), Value::Document(value_doc.clone()));
-    }
+pub unsafe extern "C" fn PLDB_doc_set_doc(doc: *mut Rc<Document>, key: *const c_char, value: *const Rc<Document>) -> c_int {
+    let local_doc = doc.as_mut().unwrap();
+    let key_str = CStr::from_ptr(key);
+    let utf8_key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
+    let local_doc_mut = Rc::get_mut(local_doc).unwrap();
+    let value_doc = value.as_ref().unwrap();
+    local_doc_mut.insert(utf8_key.into(), Value::Document(value_doc.clone()));
     0
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_doc_set_arr(doc: *mut Rc<Document>, key: *const c_char, value: *const Rc<Array>) -> c_int {
-    unsafe {
-        let local_doc = doc.as_mut().unwrap();
-        let key_str = CStr::from_ptr(key);
-        let utf8_key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
-        let local_doc_mut = Rc::get_mut(local_doc).unwrap();
-        let value_arr = value.as_ref().unwrap();
-        local_doc_mut.insert(utf8_key.into(), Value::Array(value_arr.clone()));
-    }
+pub unsafe extern "C" fn PLDB_doc_set_arr(doc: *mut Rc<Document>, key: *const c_char, value: *const Rc<Array>) -> c_int {
+    let local_doc = doc.as_mut().unwrap();
+    let key_str = CStr::from_ptr(key);
+    let utf8_key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
+    let local_doc_mut = Rc::get_mut(local_doc).unwrap();
+    let value_arr = value.as_ref().unwrap();
+    local_doc_mut.insert(utf8_key.into(), Value::Array(value_arr.clone()));
     0
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_doc_set_object_id(doc: *mut Rc<Document>, key: *const c_char, value: *const ObjectId) -> c_int {
-    unsafe {
-        let local_doc = doc.as_mut().unwrap();
-        let key_str = CStr::from_ptr(key);
-        let utf8_key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
-        let local_doc_mut = Rc::get_mut(local_doc).unwrap();
-        let value_oid = value.as_ref().unwrap();
-        local_doc_mut.insert(utf8_key.into(), Value::ObjectId(Rc::new(value_oid.clone())));
-    }
+pub unsafe extern "C" fn PLDB_doc_set_object_id(doc: *mut Rc<Document>, key: *const c_char, value: *const ObjectId) -> c_int {
+    let local_doc = doc.as_mut().unwrap();
+    let key_str = CStr::from_ptr(key);
+    let utf8_key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
+    let local_doc_mut = Rc::get_mut(local_doc).unwrap();
+    let value_oid = value.as_ref().unwrap();
+    local_doc_mut.insert(utf8_key.into(), Value::ObjectId(Rc::new(value_oid.clone())));
     0
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_doc_set_UTCDateTime(doc: *mut Rc<Document>, key: *const c_char, value: i64) -> c_int {
-    unsafe {
-        let local_doc = doc.as_mut().unwrap();
-        let key_str = CStr::from_ptr(key);
-        let utf8_key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
-        let local_doc_mut = Rc::get_mut(local_doc).unwrap();
-        local_doc_mut.insert(utf8_key.into(), Value::UTCDateTime(Rc::new(UTCDateTime::new(value as u64))));
-    }
+pub unsafe extern "C" fn PLDB_doc_set_UTCDateTime(doc: *mut Rc<Document>, key: *const c_char, value: i64) -> c_int {
+    let local_doc = doc.as_mut().unwrap();
+    let key_str = CStr::from_ptr(key);
+    let utf8_key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
+    let local_doc_mut = Rc::get_mut(local_doc).unwrap();
+    local_doc_mut.insert(utf8_key.into(), Value::UTCDateTime(Rc::new(UTCDateTime::new(value as u64))));
     0
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_doc_get(doc: *mut Rc<Document>, key: *const c_char, result: *mut *mut Value) -> c_int {
-    unsafe {
-        let local_doc = doc.as_mut().unwrap();
-        let key_str = CStr::from_ptr(key);
-        let utf8_key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
-        let get_result = local_doc.get(utf8_key);
-        if let Some(value) = get_result {
-            let out_box = Box::new(value.clone());
-            result.write(Box::into_raw(out_box));
-            return 1;
-        }
-        return 0;
+pub unsafe extern "C" fn PLDB_doc_get(doc: *mut Rc<Document>, key: *const c_char, result: *mut *mut Value) -> c_int {
+    let local_doc = doc.as_mut().unwrap();
+    let key_str = CStr::from_ptr(key);
+    let utf8_key = try_read_utf8!(key_str.to_str(), PLDB_error_code());
+    let get_result = local_doc.get(utf8_key);
+    if let Some(value) = get_result {
+        let out_box = Box::new(value.clone());
+        result.write(Box::into_raw(out_box));
+        return 1;
     }
+    return 0;
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_doc_len(doc: *mut Rc<Document>) -> c_int {
-    unsafe {
-        let local_doc = doc.as_mut().unwrap();
-        let len = local_doc.len();
-        len as c_int
-    }
+pub unsafe extern "C" fn PLDB_doc_len(doc: *mut Rc<Document>) -> c_int {
+    let local_doc = doc.as_mut().unwrap();
+    let len = local_doc.len();
+    len as c_int
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_doc_iter(doc: *mut Rc<Document>) -> *mut Iter<'static, String, Value> {
-    unsafe {
-        let local_doc = doc.as_mut().unwrap();
-        let iter = local_doc.iter();
-        Box::into_raw(Box::new(iter))
-    }
+pub unsafe extern "C" fn PLDB_doc_iter(doc: *mut Rc<Document>) -> *mut Iter<'static, String, Value> {
+    let local_doc = doc.as_mut().unwrap();
+    let iter = local_doc.iter();
+    Box::into_raw(Box::new(iter))
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_doc_iter_next(iter: *mut Iter<'static, String, Value>,
+pub unsafe extern "C" fn PLDB_doc_iter_next(iter: *mut Iter<'static, String, Value>,
                                      key_buffer: *mut c_char, key_buffer_size: c_uint, out_val: *mut *mut Value) -> c_int {
 
-    unsafe {
-        let local_iter = iter.as_mut().unwrap();
-        let tuple = local_iter.next();
-        match tuple {
-            Some((key, value)) => {
-                let key_len = key.len();
-                if key_len > (key_buffer_size as usize) {
-                    set_global_error(DbErr::BufferNotEnough(key_len));
-                    return PLDB_error_code();
-                }
-                let real_size = std::cmp::min(key_len, key_buffer_size as usize);
-
-                let cstr = CString::new(key.clone()).unwrap();
-                cstr.as_ptr().copy_to_nonoverlapping(key_buffer, real_size);
-
-                let boxed_value = Box::new(value.clone());
-                out_val.write(Box::into_raw(boxed_value));
-                real_size as c_int
+    let local_iter = iter.as_mut().unwrap();
+    let tuple = local_iter.next();
+    match tuple {
+        Some((key, value)) => {
+            let key_len = key.len();
+            if key_len > (key_buffer_size as usize) {
+                set_global_error(DbErr::BufferNotEnough(key_len));
+                return PLDB_error_code();
             }
+            let real_size = std::cmp::min(key_len, key_buffer_size as usize);
 
-            None => {
-                0
-            }
+            let cstr = CString::new(key.clone()).unwrap();
+            cstr.as_ptr().copy_to_nonoverlapping(key_buffer, real_size);
 
+            let boxed_value = Box::new(value.clone());
+            out_val.write(Box::into_raw(boxed_value));
+            real_size as c_int
         }
+
+        None => {
+            0
+        }
+
     }
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_free_doc_iter(iter: *mut Iter<'static, String, Value>) {
-    unsafe {
-        let _ptr = Box::from_raw(iter);
-    }
+pub unsafe extern "C" fn PLDB_free_doc_iter(iter: *mut Iter<'static, String, Value>) {
+    let _ptr = Box::from_raw(iter);
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_free_doc(doc: *mut Rc<Document>) {
-    unsafe {
-        let _ptr = Box::from_raw(doc);
-    }
+pub unsafe extern "C" fn PLDB_free_doc(doc: *mut Rc<Document>) {
+    let _ptr = Box::from_raw(doc);
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_mk_UTCDateTime(time: i64) -> *mut UTCDateTime {
+pub unsafe extern "C" fn PLDB_mk_UTCDateTime(time: i64) -> *mut UTCDateTime {
     let datetime = if time >= 0 {
         UTCDateTime::new(time as u64)
     } else {
@@ -951,51 +842,43 @@ pub extern "C" fn PLDB_mk_UTCDateTime(time: i64) -> *mut UTCDateTime {
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_UTCDateTime_get_timestamp(dt: *const UTCDateTime) -> i64 {
-    unsafe {
-        let dt = dt.as_ref().unwrap();
-        dt.timestamp() as i64
-    }
+pub unsafe extern "C" fn PLDB_UTCDateTime_get_timestamp(dt: *const UTCDateTime) -> i64 {
+    let dt = dt.as_ref().unwrap();
+    dt.timestamp() as i64
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_free_UTCDateTime(dt: *mut UTCDateTime) {
-    unsafe {
-        let _ = Box::from_raw(dt);
-    }
+pub unsafe extern "C" fn PLDB_free_UTCDateTime(dt: *mut UTCDateTime) {
+    let _ = Box::from_raw(dt);
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_mk_object_id(db: *mut DbContext) -> *mut ObjectId {
-    let rust_db = unsafe { db.as_mut().unwrap() };
+pub unsafe extern "C" fn PLDB_mk_object_id(db: *mut DbContext) -> *mut ObjectId {
+    let rust_db = db.as_mut().unwrap();
     let oid = rust_db.object_id_maker().mk_object_id();
     let oid = Box::new(oid);
     Box::into_raw(oid)
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_free_object_id(oid: *mut ObjectId) {
-    unsafe {
-        let _ptr = Box::from_raw(oid);
-    }
+pub unsafe extern "C" fn PLDB_free_object_id(oid: *mut ObjectId) {
+    let _ptr = Box::from_raw(oid);
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_object_id_to_hex(oid: *const ObjectId, buffer: *mut c_char, buffer_size: c_uint) -> c_int {
-    unsafe  {
-        let rust_oid = oid.as_ref().unwrap();
-        let oid_hex = rust_oid.to_hex();
-        let size = oid_hex.len();
-        let cstr = CString::new(oid_hex).unwrap();
-        let real_size = std::cmp::min(size, buffer_size as usize);
-        cstr.as_ptr().copy_to_nonoverlapping(buffer, real_size);
-        real_size as c_int
-    }
+pub unsafe extern "C" fn PLDB_object_id_to_hex(oid: *const ObjectId, buffer: *mut c_char, buffer_size: c_uint) -> c_int {
+    let rust_oid = oid.as_ref().unwrap();
+    let oid_hex = rust_oid.to_hex();
+    let size = oid_hex.len();
+    let cstr = CString::new(oid_hex).unwrap();
+    let real_size = std::cmp::min(size, buffer_size as usize);
+    cstr.as_ptr().copy_to_nonoverlapping(buffer, real_size);
+    real_size as c_int
 }
 
 #[no_mangle]
-pub extern "C" fn PLDB_free_value(val: *mut Value) {
-    let _val = unsafe { Box::from_raw(val) };
+pub unsafe extern "C" fn PLDB_free_value(val: *mut Value) {
+    let _val = Box::from_raw(val);
 }
 
 fn error_code_of_db_err(err: &DbErr) -> i32 {
